🔒 fix(device): derive worktree path server-side & thread workspaceId
Address review findings on the remote create-worktree RPC:

- P1: `device.addGitWorktree` no longer accepts `worktreePath` from the
  request. The target is re-derived on the trusted server from `path` +
  `branch` (a `<repo>-<branch>` sibling; branch folded to `-` so it can't
  traverse), so a crafted authenticated web call can't ask a remote device
  to check out at an arbitrary absolute path. `deriveWorktreePath` is
  lifted to `@lobechat/types` as the single source shared by the renderer
  (preview + local IPC) and the server (derivation).
- P2: thread `ctx.workspaceId` through `addGitWorktree` (and the sibling
  `removeGitWorktree`, which had the same gap) so workspace-scoped sessions
  route to the workspace-owned device pool instead of the personal one.

Local desktop IPC is unchanged — this machine owns its filesystem, so the
renderer-derived path is still forwarded there.
