+++ packages/types/src/device.ts
@@ -138,6 +138,33 @@ export const getWorkingDirEffectivePath = (
   return entry.git?.activeWorktree || entry.path;
 };
 
+/**
+ * Derive the target directory for a new worktree: a sibling of the source repo
+ * named `<repoName>-<branch>` (e.g. `/code/lobehub` + `feat/x` →
+ * `/code/lobehub-feat-x`), matching the convention agents already use for their
+ * linked worktrees. Preserves the source path's separator so Windows paths stay
+ * intact, and folds ref-illegal characters in the branch to `-` for the folder.
+ *
+ * Shared by the renderer (path preview + local IPC call) and the server
+ * (`device.addGitWorktree`), which re-derives the target from the trusted
+ * `path` + `branch` rather than trusting a client-supplied absolute path — so a
+ * crafted web request can't ask a remote device to check out at an arbitrary
+ * location. Both callers must derive identically, hence the single source here.
+ */
+export const deriveWorktreePath = (sourcePath: string, branch: string): string => {
+  const sep = sourcePath.includes('\\') && !sourcePath.includes('/') ? '\\' : '/';
+  const trimmed = sourcePath.replace(/[/\\]+$/, '');
+  const cut = Math.max(trimmed.lastIndexOf('/'), trimmed.lastIndexOf('\\'));
+  const parent = cut >= 0 ? trimmed.slice(0, cut) : '';
+  const repoName = (cut >= 0 ? trimmed.slice(cut + 1) : trimmed) || 'repo';
+  const suffix = branch
+    .trim()
+    .replaceAll(/[\s~^:?*[\]\\/]+/g, '-')
+    .replaceAll(/^-+|-+$/g, '');
+  const folder = suffix ? `${repoName}-${suffix}` : repoName;
+  return parent ? `${parent}${sep}${folder}` : folder;
+};
+
 export interface WorkingDirEntry extends WorkingDirConfig {
   /**
    * Cached "workspace init" scan of this directory (AGENTS.md + project skills).
